feat(onChange): call with all downshift state and helpers
**What**: calls `onStateChange`, `onChange`, and `onUserAction` with all
the downshift state and helpers.

**Why**: Make it so people don't need to use the downshift reference.
Closes #155

**How**: Instead of calling `this.getState()`, call
`this.getDownshiftStateAndHelpers()`.
